feat(rust): deploy our own map plugin, remove RustadminOnline
RustadminOnline failed to compile after a Rust update and took the panel's map
with it. RustdMap does the one thing the panel needed from it.

The old plugin is removed explicitly rather than merely superseded: left on
disk it keeps failing to compile and shows as a permanent error in the plugins
listing. Both the weekly and monthly servers already had a "Copy oxide
plugins" block deploying RustadminOnline, so both are updated and cleaned up.
